updated about box and minor port fixes

+++ action_plugins/macro/__init__.py
@@ -495,8 +495,8 @@ class MacroActionEditor(QtWidgets.QWidget):
         geom = root.geometry()
 
         self.button_press_dialog.setGeometry(
-            geom.x() + geom.width() / 2 - 150,
-            geom.y() + geom.height() / 2 - 75,
+            int(geom.x() + geom.width() / 2 - 150),
+            int(geom.y() + geom.height() / 2 - 75),
             300,
             150
         )
