+++ action_plugins_old/map_to_mouse/__init__.py
@@ -300,8 +300,8 @@ class MapToMouseWidget(gremlin.ui.input_item.AbstractActionWidget):
         geom = root.geometry()
 
         self.button_press_dialog.setGeometry(
-            geom.x() + geom.width() / 2 - 150,
-            geom.y() + geom.height() / 2 - 75,
+            int(geom.x() + geom.width() / 2 - 150),
+            int(geom.y() + geom.height() / 2 - 75),
             300,
             150
         )
